+++ src/Drivers/ManagerContract.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 /**
- * @author enea dhack <hello@enea.io>
+ * @author enea dhack <me@enea.io>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Enea\Authorization\Drivers;
 
-use Enea\Authorization\Contracts\GrantableOwner;
+use Enea\Authorization\Contracts\Owner;
 use Enea\Authorization\Contracts\PermissionsOwner;
 use Enea\Authorization\Contracts\RolesOwner;
 use Illuminate\Support\Collection;
@@ -22,5 +22,5 @@ interface ManagerContract
 
     public function roles(RolesOwner $owner): Collection;
 
-    public function forget(GrantableOwner $owner): void;
+    public function forget(Owner $owner): void;
 }