+++ src/Operators/Granter.php
@@ -3,14 +3,17 @@
 declare(strict_types=1);
 
 /**
- * Created on 12/02/18 by enea dhack.
+ * @author enea dhack <me@enea.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
 namespace Enea\Authorization\Operators;
 
 use Closure;
 use Enea\Authorization\Contracts\{
-    Grantable, GrantableOwner, PermissionsOwner, RolesOwner
+    Grantable, Owner, PermissionContract, PermissionsOwner, RolesOwner
 };
 use Enea\Authorization\Events\Granted;
 use Enea\Authorization\Exceptions\AuthorizationNotGrantedException;
@@ -20,9 +23,13 @@ use Illuminate\Support\Collection;
 
 class Granter extends Operator
 {
+    use WithDeniablePermission;
+
     public function permissions(PermissionsOwner $owner, Collection $permissions): void
     {
-        $this->grantTo($owner->permissions())($permissions);
+        $granted = $this->getPermissions($owner, $permissions);
+        $granted->each($this->denialStatus($owner, false));
+        $this->grantTo($owner->permissions())($this->except($granted, $permissions));
         $this->dispatchGrantedEvent($owner, $permissions);
     }
 
@@ -55,8 +62,18 @@ class Granter extends Operator
         }
     }
 
-    private function dispatchGrantedEvent(GrantableOwner $owner, Collection $grantableCollection): void
+    private function dispatchGrantedEvent(Owner $owner, Collection $grantableCollection): void
     {
         $this->dispatchEvent(new Granted($owner, $grantableCollection));
+    }
+
+    protected function isModifiable(PermissionContract $permission): bool
+    {
+        return $permission->pivot->isDenied();
+    }
+
+    protected function throwException(Grantable $grantable): void
+    {
+        throw new AuthorizationNotGrantedException($grantable);
     }
 }